requirements.txt
# Traveco Transport Forecasting Project
# Python dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Forecasting libraries
prophet>=1.1.4
statsmodels>=0.14.0
xgboost>=2.0.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Excel file support
openpyxl>=3.1.0
pyxlsb>=1.0.10

# Additional utilities
python-dateutil>=2.8.2
pyarrow>=14.0.0        # Parquet file support

# Production Application (NEW)
streamlit>=1.29.0      # Lightweight dashboard
click>=8.1.0           # CLI framework
rich>=13.7.0           # Terminal formatting
python-dotenv>=1.0.0   # Environment variables

# Testing (NEW)
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Logging
loguru>=0.7.2

# Optional: Development
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
